requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework==3.13.1
numpy==1.23.1
pymemcache==3.5.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
